requirements.txt
Flask
Flask-Login
sqlite3
discord.py
werkzeug